Renamed `src/app` to `src/application`
